refactor: application 단위에서 repository 주입
