--- app/src/main/java/woowacourse/shopping/data/history/HistoryDatabase.kt
@@ -1,27 +1,0 @@
-package woowacourse.shopping.data.history
-
-import android.content.Context
-import androidx.room.Database
-import androidx.room.Room
-import androidx.room.RoomDatabase
-
-@Database(entities = [HistoryEntity::class], version = 1)
-abstract class HistoryDatabase : RoomDatabase() {
-    abstract fun historyDao(): HistoryDao
-
-    companion object {
-        @Volatile
-        private var instance: HistoryDatabase? = null
-
-        fun getDatabase(context: Context): HistoryDatabase =
-            instance ?: synchronized(this) {
-                Room
-                    .databaseBuilder(
-                        context.applicationContext,
-                        HistoryDatabase::class.java,
-                        "history_db",
-                    ).build()
-                    .also { instance = it }
-            }
-    }
-}
